🚚 api: move invigilator kit pdf to src
