+++ api/src/certification/session/infrastructure/utils/pdf/files/invigilator-kit_template.pdf
