+++ api/src/certification/session/infrastructure/utils/pdf/files/kit-surveillant_template.pdf
