+++ api/tests/certification/session/integration/infrastructure/utils/pdf/invigilator-kit-with-long-labels_expected-v3.pdf
