+++ api/tests/certification/session/integration/infrastructure/utils/pdf/invigilator-kit_expected-v3.pdf
