+++ api/tests/certification/session/integration/infrastructure/utils/pdf/invigilator-kit_expected.pdf
